# omcgo/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OMC API
  description: Operations Management and Control for Base Stations
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/config/templates:
    get:
      tags: [Templates]
      summary: List configuration templates
      operationId: listConfigTemplates
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated template list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConfigTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Templates]
      summary: Create a configuration template
      operationId: createConfigTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigTemplateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/config/templates/{id}:
    get:
      tags: [Templates]
      summary: Get configuration template by ID
      operationId: getConfigTemplate
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Templates]
      summary: Update a configuration template
      operationId: updateConfigTemplate
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConfigTemplateRequest"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Templates]
      summary: Delete a configuration template
      operationId: deleteConfigTemplate
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "204":
          description: Template deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================================================================
  # ConfigSync
  # ===========================================================================
...
  /api/v1/northbound/push/targets:
    get:
      tags: [Northbound]
      summary: List push targets
      operationId: listPushTargets
      responses:
        "200":
          description: Push target list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PushTarget"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Northbound]
      summary: Create a push target
      operationId: createPushTarget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePushTargetRequest"
      responses:
        "201":
          description: Push target created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTarget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/northbound/push/targets/{id}:
    delete:
      tags: [Northbound]
      summary: Delete a push target
      operationId: deletePushTarget
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "204":
          description: Push target deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/northbound/sync/full:
    get:
      tags: [Northbound]
      summary: Get full northbound sync data
      operationId: getFullSync
      responses:
        "200":
          description: Full sync data
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/northbound/sync/incremental:
    get:
      tags: [Northbound]
      summary: Get incremental northbound sync data
      operationId: getIncrementalSync
      responses:
        "200":
          description: Incremental sync data
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/northbound/export/pm:
    post:
      tags: [Northbound]
      summary: Export PM data
      operationId: exportPMData
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                format:
                  type: string
                  enum: [json, csv, xml]
      responses:
        "200":
          description: PM data exported
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/northbound/export/alarms:
    post:
      tags: [Northbound]
      summary: Export alarm data
      operationId: exportAlarmData
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                format:
                  type: string
                  enum: [json, csv, xml]
      responses:
        "200":
          description: Alarm data exported
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/northbound/export/config/{deviceId}:
    get:
      tags: [Northbound]
      summary: Export device configuration
      operationId: exportDeviceConfig
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Device configuration export
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================================================================
  # Admin - Users
  # ===========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    UUIDPathParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
        request_id:
          type: string
          description: Request tracking ID
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of records
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20

    # ---- Auth ----
...
    ConfigTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        carrier:
          type: string
        technology:
          type: string
        content:
          type: object
          description: Template parameter tree
          additionalProperties: true
        version:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateConfigTemplateRequest:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        description:
          type: string
        carrier:
          type: string
        technology:
          type: string
        content:
          type: object
          additionalProperties: true
        version:
          type: string
    UpdateConfigTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: object
          additionalProperties: true
        version:
          type: string

    # ---- Provisioning ----
...
    PushTarget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [http, kafka, mqtt]
        url:
          type: string
        auth_type:
          type: string
          enum: [none, basic, bearer, api_key]
        enabled:
          type: boolean
        data_types:
          type: array
          items:
            type: string
            enum: [alarm, pm, config, event]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePushTargetRequest:
      type: object
      required: [name, type, url]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [http, kafka, mqtt]
        url:
          type: string
        auth_type:
          type: string
          enum: [none, basic, bearer, api_key]
        auth_config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
          default: true
        data_types:
          type: array
          items:
            type: string
            enum: [alarm, pm, config, event]

    # ---- User / Role / Permission ----
...